# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HealApp API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/stats/chart": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get diary chart data",
                "description": "Retrieve diary entries for a specific patient and key (e.g., temperature) over a time period. Returns data points for chart visualization.",
                "operationId": "07c105ac09a89ec83da1d29c0e3051e3",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "query",
                        "description": "Patient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Diary entry key (e.g., 'temperature', 'blood_pressure', 'mood')",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "temperature"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Time period for data",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "7_days",
                            "enum": [
                                "7_days",
                                "30_days"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chart data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "patient_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "temperature"
                                        },
                                        "period": {
                                            "type": "string",
                                            "example": "7_days"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "recorded_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T08:00:00.000000Z"
                                                    },
                                                    "value": {
                                                        "description": "JSON value from diary entry",
                                                        "type": "object",
                                                        "example": {
                                                            "value": "36.6"
                                                        }
                                                    },
                                                    "notes": {
                                                        "type": "string",
                                                        "example": "Normal temperature",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - missing required parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "patient_id and key parameters are required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have permission to access this patient."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Patient] {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/stats/tasks": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get task summary statistics",
                "description": "Retrieve task statistics for a patient on a specific date. Returns counts and completion rate.",
                "operationId": "c4e8dabfe5901fb914477cb1abbcaa5a",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "query",
                        "description": "Patient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date for statistics (YYYY-MM-DD). Defaults to today.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task summary retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "patient_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-01-01"
                                        },
                                        "total": {
                                            "description": "Total number of tasks",
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "completed": {
                                            "description": "Number of completed tasks",
                                            "type": "integer",
                                            "example": 7
                                        },
                                        "missed": {
                                            "description": "Number of missed tasks",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "pending": {
                                            "description": "Number of pending tasks",
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "completion_rate": {
                                            "description": "Completion rate as percentage (0-100)",
                                            "type": "number",
                                            "format": "float",
                                            "example": 70
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - patient_id is required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "patient_id parameter is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have permission to access this patient."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Patient] {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}